Add support for augmented security rules
This support was already added to the `azurerm_network_security_rule` resource in PR #692, but it wasn’t added to the `azurerm_network_security_group` resource at that time.

It did needed a bit of additional logic to make it work properly in the `azurerm_network_security_group` resource as well, but it all worked out nicely.

All related unit- and acceptance tests were successfully run before creating this PR.