Send provider version in user agent
This fixes user agents to conform to the RFC, it also takes advantage
of a new build flag populating the binary version of the provider.

The version sent in the user agent without the flag will be
`terraform-provider-azurerm/dev` and for acceptance tests through
`make testacc` `terraform-provider-azurerm/acc`. The binary should
send a product in the user agent in the form of
`terraform-provider-azurerm/1.2.3`.