Support app_command_line in App Service site_config
App Service supports setting a command line to use with Linux
applications (e.g. to override the default in the container image).
Expose this configuration as `app_command_line` in `site_config` for
`azurerm_app_service`.

N.B. This is also referred to as "startup file" by some sources,
including the `az` Azure CLI and the Azure Portal.

References:

- https://docs.microsoft.com/en-us/rest/api/appservice/webapps/createorupdate#siteconfig
- https://docs.microsoft.com/en-us/cli/azure/webapp?view=azure-cli-latest#az-webapp-create