Azure batch expose keys and account url (#3071)
This PR exposes the primary access key, secondary access key and account endpoint in the `azurerm_batch_account` resource and data_source.  The documentation has also been updated, with a note that the keys are only available when the batch accounts `pool_allocation_mode` is `BatchService`.  If the allocation mode is `UserSubscription`, terraform will throw an error.