Added note about `create_option` (#4348)
For the `storage_data_disk` block, there may be some confusion as to how to properly use the `create_option` parameter. In order to successfully deploy a data disk while using a template that does not initially write data to a data disk, the `Empty` parameter value must be specified. This was not clear during my initial attempts to deploy. Thanks.