r/kubernetes_cluster: fix updating cluster with autoscaling enabled
Currently, on every update, agentPoolProfiles are build from scratch
based on pool definition from Terraform. With autoscaled clusters, this
attempts to remove 'Count' property, which triggers manual scaling
action, which is forbidden for clusters with autoscaling enabled. That
breaks any updates to the cluster, including adding tags or updating
kubernetes version.

With this commit, we always try to fetch definition of the cluster from
the API and if we update, we only modify profile parameters using Terraform
configuration, rather than building profiles from scratch. This makes
sure that all parameters set by API will be preserved.

To enforce this behavior, we should fail when we are updating the
resource, but we are not able to fetch it from the API.

'expandKubernetesClusterAgentPoolProfiles' function now accepts profiles
slice, which it will work on and returns modified copy, rather than
newly build profiles slice.

Closes #4075

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>