Maintain Agent Pool Profile order in AKS
This commit takes the names of the agent pool profiles as defined in the
terraform configuration and uses them to sort the list of profiles
returend from the Azure API. Since our config uses `schema.TypeList`,
the order of the config should be respected. However, it appears the
Azure API returns the profiles in alphabetical order. This commit fixes
this.