storage_account: Fix Azure Germany ATP matching
Although the error message returned from Azure Germany Cloud is
precisely "The resource namespace 'Microsoft.Security' is invalid."
the error message we are seeing in `r/storage_account` has other
contents added on it, so we cannot do an equals check directly.
The ATP fix was being pushed as v1.36.0, but this change provides
the actual resolution.

Ref #4563, #4564.