Fix issue #584 add option to set AI daily cap and hit daily cap… (#5480)
This PR should fix #584, so we can set now a daily cap for Applications Insight in GB. Beside setting a daily cap for Applications Insight it's also possible now to disable notifications if daily cap is hit.

Compared to other resources this configuration is separated to it's own billing feature api. Which requires to update billing feature api once Application Insight was created.