provider: making the `features` block required
This might seem like an odd choice - but since this block contains switches which can alter
the behaviour of certain resources (including destructive behaviour) - we intentionally want
to highlight the presence of this block.

Whilst the `features` block itself is required - the inner elements are not, thus is a user
wishes to use the default features they can specify:

```
provider "azurerm" {
  version = "=2.0.0"

  features {}
}
```

However it's also possible to override certain features, for example:

```
provider "azurerm" {
  version = "=2.0.0"

  features {
    virtual_machine {
      delete_os_disk_on_deletion = false
    }
  }
}
```

Since this block is Required this'll also mean that Provider blocks need to be threaded through
to modules (which they should be being today, but this'll make this explicit).