Couple default_node_pool OrchestratorVersion to cluster k8s_version
Per #5541, currently AKS node pool versions can never be updated. This
occurred due to a change in behavior in new ARM behavior now that AKS
clusters can have multiple agent pools.

There's a more involved fix in the discussion on that issue which involves
exposing OrchestratorVersion as a settable attribute on agent_pool_profiles,
but this change focuses on recovering the old behavior.