tests: clarifying the features block is required going forward
The temporary patch in the provider/test suite allows for us to gradually add provider
blocks to the test suite, rather than doing it in a big-bang (and breaking open PR's).

Unfortunately this workaround was unintentionally available in release binaries, as such
this commit changes this patch to only work for test providers rather than released versions
since this isn't intended to be used by users.

Instead we require that Provider blocks and Feature blocks are both specified to be explicit
about which instance of a provider is being used when provisioning resources.