Added loadbalancer name-addresses map attribute
This adds a new attribute to get a name-private_ip_address mapping
of the frontend_ip_configurations, defined on the loadbalancer.
Helpful, if you have manny frontend_ip_configurations, created with a
dynamic loop and need these mapping for further steps or outputs,
otherwise you just rely on the list order.