azurerm_application_gateway: can combine include_path with target_url
Redirecting say http://foo/path?query to http://bar/
include_path controls whether it will go to
	http://bar/path
or
	http://bar/

This is a perfectly valid option in a redirect_configuration in
combination with target_url. Currently the code gives an error if you
use them in combination and omits it from the request if target_url is
set.

Omitting include_path has strange results in the Azure Portal
(no radio button selected) and "null" defaults to true whereas the
documentation says include_path should default to false.

This change makes the behavior consistent with the documentation.