Issue #6059 - subscription rights for key_vault
This commit fixes issue #6509, in which some subscription level
rights are required in order to create a key_vault. This occurred
due to the provider always looking for a soft deleted key vault
prior to checking whether the feature
KeyVault.RecoverSoftDeletedKeyVaults was set to true. Thus event
setting the value to false did not stop it searching for the soft
deletions.

The rights it was requiring were to perform action
'Microsoft.KeyVault/locations/deletedVaults/read' over scope
'/subscriptions/xxxxxx'