azurerm_cognitive_account - support for the `qna_runtime_endpoi… (#5778)
This commit creates a new optional qna_runtime_endpoint property at the top level of the cognitive account resource

Previously if the kind of a cognitive service account was set to QnAMaker an error would be thrown from the Microsoft API (see issue for more details).
This issue was blocked due to missing functionality in the Azure SDK however support to allow the Account Properties and Account API Properties to be set was added in V25. This change allows the qna_runtime_endpoint to be set and will allow QnAMaker cognitive service accounts to be created via Terraform.

I have added 2 Acceptance Tests for setting the qna_runtime_endpoint with QnAMaker set as the kind, 1 as a new resource and the 2nd with the url updated. I tried to add an acceptance test for setting the qna_runtime_endpoint on a non QnAMaker kind cognitive service account i.e. Face however the Microsoft API ignores this property when it is supplied so it is not returned in the response. As a result the tests failed with the following error ImportStateVerify attributes not equivalent. I decided remove this test.

The website docs have been updated to match the current functionality.

Fixes #4338

I am still quite new to Go and this is my first commit to the Terraform Providers so any feedback will be greatly appreciated