New resource `azurerm_policy_remediation` (#5746)
This is a revised version of PR 5277.

A azurerm_policy_remediation is used along with a azurerm_policy_assignment to remediate a policy.
A policy can be assigned in 4 different scopes: Subscription, Resource Group, Resource and Management Group. This remediation resource supports all of them using a scope field similar to the azurerm_policy_assignment.

Since the id of a management group (like /providers/Microsoft.Management/managementGroups/00000000-0000-0000-0000-000000000000) is not a valid azure resource id, passing it to azure.ParseResourceId will throw errors, I wrote some additional ID parsing functions to deal with this kind of situations.

And additionally, since all of the ids in a remediation will be converted to lower cases, I am using case diff suppression everywhere.