azurerm_management_group - deprecated and rename `group_id` to… (#6276)
The term group_id in management group is sometimes confusing with the UUID of the mgmt group or a full ID of a mgmt group. And a mgmt group can have a human-readable string as its group_id. Therefore group_id is not quite a good term for this.

I introduced a new attribute name for the resource and data source azurerm_management_group to provide more friendly usage of management group and reduce the confusion.