r/kubernetes_cluster: making the `aci_connector_linux` & `http_application_routing` blocks cloud specific
The blocks `aci_connector_linux` and `http_application_routing` are
not supported in Azure China or Azure US Government Clouds. As such
this commit filters them out if they're disabled (since the API does
not allow them to exist in the payload, even if disabled) and returns
an error if attempting to use these add-ons in this region.