Add linux_syslog toggle in workspace
There is a dedicated workspace-level resource called `linux_syslog_collection`,
which has only a `state` indicating whether `linux_syslog` DS is applied
to all the machines in this workspace. Rather than creating a
corresponding resource for the "collection", making it as a workspace
property seems more intuitive in context of terraform.