New resource DF Self-host Integration Runtime
This resource type supports but a standalone Self-hosted and linking via rbac_authorization. The
key_authorization has no documentation other than the REST interface.  The Powershell Cmdlet does
not make use of it.  So it has been left out.

A full test is setup and necessary for the linked environment to work.

An example and documetnation is included to make the linked setup easier to understand.  People
may also reference the test case.

Fixing issue #6503.