tooling: Provider Defined Builds
This commit introduces Provider Defined Builds - which allows us to define
the Build Configuration used for Acceptance Testing as code in this repository.

As a part of this we dynamically generate Build Configurations based on the
Services which exist in this repository - thus adding a new Build Configuration
will automatically (once merged) add a new matching Build Configuration.