tool: adding a check for gradually deprecated functions
Go's linter immediately flags usages of functions which are marked as
Deprecated - which makes sense when there's 10 usages, but less so when
there's hundreds of usages and you want to remove things gradually.

This PR introduces a check for functions which will be deprecated in the
future, since they're no longer required - namely the feature flag
"RequireResourcesToBeImported" which is now enabled by default.

Over time we'll work through the existing codebase and remove usages of
this function - but we intentionally don't want to do this in a single
sweep as to not break open PR's (since this defaulted to "enabled" there's
not much harm in this code kicking around for a while)