azuerm_function_app - correctly set kind when os_type is `linux` (#7140)
There is a typo in the code where osType is "Linux" instead of "linux". The format does not match with checks and other parts of the code such as in row 354.

Current behaviour causes the function app to be generated as Windows despite the value of os_type. This fixes #6931 for example.