azurerm_marketplace_agreement: fix import issue
The import error message is incorrect, as this resource is using the
ID of `agreement` as its ID, it should use that ID to imform user to
import (while currently it is using the ID of `term` in the error
message).

In the meanwhile, modify some other log messages to convey the correct
intention.

Besides, fix a typo in another PR which fixes the import ID in the
document.