Update `azurerm_shared_image` - flip `generalized` to `specialized` (#7525)
Since the changes on azurerm_shared_image has not been released yet, this should not be a breaking change. I suppose it could make more sense if we name this new attribute specialized to emphasize when we are assigning a non-zero value to it, the image is specialized. Rather than previous behaviour that we assign generalized to false to indicate this is specialized.