fixes regression related to #7879
For those storage account created using old API version (at least
2018-02-01, before provider v1.29.0), the GET call of new API version
will not return the `minimumTLSVersion` in response body. So those users
will get a diff.

This PR sets the default value for `min_tls_version` in case it is
missing in response, using the same value as the default value defined
in schema.