Allow to set security center pricing tier for a particular resource type
At the moment, only Virtual Machines are set with the security center
standard or free pricing tiers when using `azurerm_security_center_subscription_pricing`.
This change adds the field `resource_type`, which allows to specify the
resource type for which we want to update the pricing tier.

The v1 security center pricing client only allows to get the pricing
tier of the default resource type (Virtual Machines) to check whether
the subscription has standard or free security center pricing tier.
However, a partial standard pricing tier (where one or more
resource types have standar tier enabled) allows for a security center
workspace to be created.

This commit changes the client to v3 and checks if any resource type
in the subscription has standard pricing tier enabled,
and if so, it allows the creation of a security center workspace.