hpc-cache: splitting out into their own service package
Ordinarily this wouldn't be worth it, but these tests are so expensive that this enables
these to be run out of band from other changes to the Storage Services, which moves far quicker.

Long-term we could look to move these to a twice-a-week/weekly run, but just ensuring these
are run daily and not for every Storage PR for now would be an easy win.