New data source: azurerm_cognitive_account (#8773)
Resolves #7765

Note that I do not use this resource and as such have not tested this functionality manually. Instead I am relying on the test coverage, and using the existing resource code, modified for the data source. If anyone uses this resource and sees anything missing in the data source that they expect to retrieve, let me know.