Update `azurerm_linux|windows_virtual_machine` and `azurerm_linux|windows_virtual_machine_scale_set` - Support managed boot diagnostics (#8917)
Fixes https://github.com/terraform-providers/terraform-provider-azurerm/issues/8319

Since add a new required attribute in the existing `boot_diagnostics` block should be considered as a breaking change, I just implement this feature like this. 
Please let me know if we could accept the other way.

Update:
I have thought of two approaches of fixing this issue, one of them is the way implemented in this PR. The other is this:
```
boot_diagnostics {
  enabled = true    # required (breaking change)
  storage_account_uri = "uri of the storage account" # optional
}
```
which would introduce a breaking change (we add a required attribute). I am not quite sure whether this is allowed. But I personally prefer this way rather than the implementation in this PR, which is more elegant and aligned with the API.