desktopvirtualization: switching to use the Insensitive option where appropriate
This allows us to workaround a bug where Core expects the Plan to 1:1 the Apply
but the Azure API returns the Resource ID's in a case-insensitive manner depending on
which API is being called.

This clearly isn't ideal, but allows for consistency on the Terraform side whilst
working around a broken API on the Azure side.