network: moving the Private DNS Zone Group parser
Whilst this looks like it should live within the Private DNS package - this resource
is ultimately managed solely within a Private Endpoint and thus likely belongs with
this resource it's nested under.

This commit also adds a missing `privateEndpoints` to the parser to match the Resource
ID being returned here - since there doesn't appear to be another option here.