Add support for service tags to ip_restriction for function_app and app_service
Service tags are now allowed in the ip restrictions. In order to support this I've weakened the field validation on ip_address field. Behaviour is that if the field does not start with a number we assume you were attempting to set a service tag, otherwise default to assuming the field contains a CIDR.

Tests and documentation will come after PR discussion.

Relates to #9380