azurerm_media_services_account - support for 'identity', 'tags' and 'storage_authentication' (#9457)
Including the support for the properties Identity and Storage Authentication in Azure Media Services Account. In order to support new fields was required to update the API version to new version 2020-05-01, but I have already validated that the Acceptance test still works. Additional take advantage of including support for tags. This PR resolves #9404