r/log_analytics_workspace: force-upgrading the resource id's
This works around an issue in the legacy state migration where the field
`resource_group` was used in place of `resource_group_name`. Whilst we
can't re-run the state migration from V0 -> V1 since those resources have
been upgraded - we can duplicate it for V1 -> V2, and patch it for new users
meaning this should work in both scenarios.