Fix App Service documentation for Linux plans (#9024)
When creating a Linux-based app service plan it was previously documented that "kind" should equal "FunctionApp" and "reserved" should equal true. This is no longer true and returns the following error: "Error: `reserved` has to be set to false when kind isn't set to `Linux`". Updated documentation to match intended usage of setting "kind" to "Linux" for linux-baseed app service plans.