Support new properties `alert_rule_template_guid` and `display_name_exclude_filter` for azurerm_sentinel_alert_rule_ms_security_incident (#9797)
The purpose of this PR is to support below two new properties for azurerm_sentinel_alert_rule_ms_security_incident.

The first property is alert_rule_template_guid. It is the id of the alert rule template which is used to create this Sentinel Scheduled Alert Rule.

The second property is display_name_exclude_filter. It is used to create incidents when the alert display name doesn't contain text from the specified list.