r/key_vault: deprecate `soft_delete_enabled` & default `soft_delete_retention_days` to 90
This commit conditionally deprecates the `soft_delete_enabled` field and removes it entirely
if we're running in 3.0 mode. This also defaults `soft_delete_retention_days` to 90 matching
the default value used in the Azure API here, based only on the docs, since this isn't returned
if it's not configured by the user.