r/azurerm_servicebus_namespace - Allowed setting messaging units capacity to 16 for Premium SKU (#10337)
Allowed the azurerm_servicebus_namespace resource to support the capacity of 16 for the number of messaging units to be set when using the Premium SKU.

Microsoft documentation (docs.microsoft.com/en-us/azure/service-bus-messaging/service-bus-premium-messaging) only shows support for a capacity of up to 8, but we have been able to request 16 due to high levels of throttling at 8. Microsoft has confirmed going up to 16 is not possible in the portal, but customers can request it via a support ticket. This change is essential for us as Terraform wants to keep reverting our capacity back to 8: