r/role_assignment: adding validation for `scope`
This commit introduces validation to the `scope` field, validating that
it's either a Management Group ID, Resource Group ID, Subscription ID
or otherwise a Resource ID - to workaround the API usability issues
identified in #9569.

This isn't perfect, but the error messages coming back from the API are
particularly unhelpful to users unfamiliar with how the API works.