fixed issue #8768 with ip restriction changes
The previous value for subnet/IP/tag was not removed when the record was going to be reused for another subnet/IP/tag. Thus we need to know which property changed so that we can determine which one of subnet/IP/tag is the one we have to keep. Passing down the ResourceData to this method is how can we pass diff information down the tree.