Customize ARM correlation request id via `ARM_CORRELATION_REQUEST_ID`
Currently, everytime Terraform initialized, it will generate a random UUID
to be used as the `x-ms-correlation-request-id` in the HTTP request sent via Go SDK.

In some cases, users might want to get that correlation request id so that
it can be used to track the E2E traces. This can be done by looking into the debug
log or using some http proxy. However, this is not an effective way in case Terraform
is running in a large scale automatically, where the debug log is disabled (for some
performance reason), also the proxy is not a good choice there. In this case, allowing
users to specify a correlation request id to be used by Terraform would be a good alternative.

This PR allows users to specify an environment variable `ARM_CORRELATION_REQUEST_ID`, which
takes higher priority than auto-generating a UUID.