fix update issue when using user assigned identity (#10994)
fix #10406

I have asked the aks service team, they say there is no such error in their codebase. The error may be thrown in ARM side or somewhere else.

according to the error message, if we set the value of identity.UserAssignedIdentities to empty, everything is ok. I think maybe somewhere the rule is strengthened, the computed value should not exist in the request body

additionally, I refactor the identity part and extract to a new file