Support Redis Replicas configuration
A Redis cluster can be configured with more than one replica per master.
See
https://docs.microsoft.com/en-us/azure/azure-cache-for-redis/cache-how-to-multi-replicas
for documentation. ReplicasPerMaster is supported in the Go SDK.

Without the ability to set replicas we are unable to set up Redis to be
fully zone redundant. One must have as many nodes aS there are zones.
THe default replica count is 1, for a total of two nodes, which would
only support two zones of redundancy. Many regions have three zones.