`azurerm_storage_account` - Workaround for Portal UI bug for Storage Account resource AllowSharedKeyAccess field value being `nil` (#11696)
AllowSharedKeyAccess can only be true due to issue 11460. If the value is nil that currently brakes the Portal UI. There is a bug in the Portal UI which interprets nil as false and per the ARM API documentation nil should be interpreted as true. This manifests itself in the Portal UI when a storage account is created by Terraform that the AllowSharedKeyAccess is Disabled when it is actually Enabled in the overview blade. This is confusing our customers and to fix this, I have added code to explicitly to set the value to true if it is nil to workaround the Portal UI bug. This is designed as a passive change/fix, meaning the change will only take effect if an existing storage account is modified in some way or if the storage account is net new. Given these changes this issue should automatically correct itself over time.

(fixes #11689)