fix error creating zone-redundant ip in central us euap (#12080)
The constant [1,2,3] is used to create zone-redundant resource, however, in location Central US EUAP, the available zones are [1,2], then it will throw an exception like the following

error: Code="InvalidAvailabilityZone" Message="The zone(s) '3' for 
resource 'Microsoft.Network/publicIPAddresses/acctestpublicip-henglu-02' is not supported. The supported zones for location 'centraluseuap' are '1,2'
It can be easily fixed by using [1,2], it can create zone-redundant resources in Central US EUAP and other locations.