Reinstate use_managed_identity=false
When service_principal_id is returned on read, the linked service is definitely not using managed id to authenticate, so this can be set false
When service_principal_id is missing on read, as per the previous commit comments, only an encrypted credential is returned, not the type.  We cannot tell if the linked service is using managed id or account key to authenticate, therefore it was correct to remove the else block containing use_managed_identity=false.