Use d.GetOk to set the parameters property of azurerm_bot_connection (#12500)
Currently, the TCs related with BotService Connection are failed. After investigated, seems currently the API behavior would fail to create bot connection when the parameters property is []. After checked, seems currently this property is optional and it would be set as [] in TF when it isn't specified in tfconfig. I assume we should use d.GetOk() for this optional property. So I submitted this PR to fix this issue.