Update BotService SDK to 2021-03-01 (#12449)
Update BotService SDK to 2021-03-01. Mainly in order to support new introduced properties and resources.

I've rerun the TCs to avoid regression issue. Below is the test result and it's same with the test result which doesn't involve this PR. So although there is failed tc in test result, but I think it isn't related with this PR.