Don't allow license_type to be set for serverless SQL databases
Serverless databases do not support a license type[0]. When we pull the
database information from the API, `licenseType` is set to `null`, which
causes plans to always have a pending change when users set
`license_type`.

This change causes the provider to throw an error before creating (or
updating) the resource.

[0]: https://docs.microsoft.com/en-us/azure/azure-sql/azure-hybrid-benefit